add better analytics, visuals are better, new functions to convert duration to formatted strings

+++ include/Human.hpp
@@ -6,7 +6,7 @@
 class Human : public Player {
 
 public:
-    Human(GameEngine *game_engine, GraphicalInterface *gui, unsigned short id);
+    Human(GameEngine *game_engine, GraphicalInterface *gui, uint8_t id);
     Human(Human const &src);
     ~Human(void);
     Human	&operator=(Human const &rhs);
